Pull out msg and state module in hackatom
