+++ contracts/hackatom/src/state.rs
@@ -1,0 +1,13 @@
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+use cosmwasm_std::CanonicalAddr;
+
+pub const CONFIG_KEY: &[u8] = b"config";
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct State {
+    pub verifier: CanonicalAddr,
+    pub beneficiary: CanonicalAddr,
+    pub funder: CanonicalAddr,
+}
